Preventing the tmp dir getting removed when run from the command line

+++ tool/bowtie_aligner.py
@@ -506,12 +506,16 @@ class bowtie2AlignerTool(Tool):  # pylint: disable=invalid-name
             for idx_file in index_files:
                 compss_delete_file(index_files[idx_file])
 
-        for fastq_file_pair in fastq_file_list:
-            compss_delete_file(gz_data_path + "/tmp/" + fastq_file_pair[0])
-            os.remove(gz_data_path + "/tmp/" + fastq_file_pair[0])
-            if "fastq2" in input_files:
-                compss_delete_file(gz_data_path + "/tmp/" + fastq_file_pair[1])
-                os.remove(gz_data_path + "/tmp/" + fastq_file_pair[1])
+        if hasattr(sys, '_run_from_cmdl') is True:
+            pass
+        else:
+            for fastq_file_pair in fastq_file_list:
+                compss_delete_file(gz_data_path + "/tmp/" + fastq_file_pair[0])
+                os.remove(gz_data_path + "/tmp/" + fastq_file_pair[0])
+                if "fastq2" in input_files:
+                    compss_delete_file(gz_data_path + "/tmp/" + fastq_file_pair[1])
+                    os.remove(gz_data_path + "/tmp/" + fastq_file_pair[1])
+
         tasks_done += 1
         logger.progress("ALIGNER", task_id=tasks_done, total=task_count)
